+++ src/Lotto.js
@@ -1,18 +1,18 @@
+import Validation from "./Validation.js";
+import { MissionUtils } from "@woowacourse/mission-utils";
+
 class Lotto {
   #numbers;
 
   constructor(numbers) {
-    this.#validate(numbers);
-    this.#numbers = numbers;
+    
+    this.#numbers = this.getNumbers();
   }
 
-  #validate(numbers) {
-    if (numbers.length !== 6) {
-      throw new Error("[ERROR] 로또 번호는 6개여야 합니다.");
-    }
+  async getNumbers () {
+    const input = await MissionUtils.Console.readLineAsync('당첨 번호를 입력해 주세요.');
+    return Validation.checkNumbers(input);
   }
 
-  // TODO: 추가 기능 구현
 }
-
 export default Lotto;
